pyproject.toml
[tool.poetry]
name = "folio-data-anonymization"
version = "0.1.0"
description = "Anonymize folio data"
authors = ["Jeremy Nelson <jpnelson@stanford.edu>", "Joshua Greben <jgreben@stanford.edu>", "Shelley Doljack <sdoljack@stanford.edu>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
apache-airflow = "^3.1.0"
faker = "^37.1.0"
jsonpath-ng = "^1.7.0"
pydantic = "^2.11.4"
psycopg2 = "^2.9.10"

[tool.poetry.group.test.dependencies]
apache-airflow = "^3.1.0"
black = "^24.10.0"
mypy = "^1.13.0"
flake8 = "^7.1.1"
pytest = "^8.3.4"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
skip-string-normalization = true
color = true

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
explicit_package_bases = true
warn_unreachable = true
pretty = true
show_error_context = true
